Split web card form lifecycle tests
